@@ -14,7 +14,7 @@ description: Fridgify - Stay cool. Stay organised.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.6.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -27,6 +27,12 @@ dependencies:
   shared_preferences: "0.5.6+3"
   flutter_datetime_picker: "^1.3.5"
   carousel_slider: "^1.4.1"
+  flutter_circular_slider: "^2.5.0"
+  sleek_circular_slider : "^1.1.0"
+  barcode_scan: "^3.0.0"
+  sortedmap: "^0.4.2+1"
+  expandable: "^4.1.3"
+  autocomplete_textfield: "^1.7.3"
   pie_chart:
     git:
        url: "https://github.com/apgapg/pie_chart.git"
